+++ packages/terra-action-footer/examples/test-examples/action-footer-container/WithActionsActionFooterContainer.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ActionFooterContainer from '../../../lib/_ActionFooterContainer';
+import Button from '../../../../terra-button/lib/Button';
+
+export default () => (
+  <ActionFooterContainer>
+    FakeLink Action
+    <Button text="Button Action" />
+  </ActionFooterContainer>
+);