+++ packages/terra-action-footer/examples/test-examples/action-footer-container/WithoutActionsActionFooterContainer.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import ActionFooterContainer from '../../../lib/_ActionFooterContainer';
+
+export default () => <ActionFooterContainer />;