+++ packages/terra-action-footer/examples/test-examples/action-footer/MultipleStartActionFooter.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ActionFooter from '../../../lib/ActionFooter';
+import Button from '../../../../terra-button/lib/Button';
+
+
+export default () => (
+  <ActionFooter
+    start={
+      <div>
+        FakeLink Start Action 1
+        <Button text="Start Action 2" />
+      </div>
+    }
+  />
+);