+++ packages/terra-avatar/tests/nightwatch/ThreeInitialsAvatar.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Avatar from '../../lib/Avatar';
+
+export default () => <Avatar id="three-initials-avatar" initials="JSR" variant="user" ariaLabel="user" />;