+++ packages/terra-site/src/examples/avatar/AvatarThreeInitials.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Avatar from '../../../../../packages/terra-avatar/src/Avatar';
+
+const AvatarThreeInitials = () => (
+  <Avatar ariaLabel="John Doe" initials="JSR" />
+);
+
+export default AvatarThreeInitials;