+++ packages/terra-site/src/examples/section-header/OpenSectionHeader.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import SectionHeaderExampleTemplate from './SectionHeaderExampleTemplate';
+
+const sectionHeaderProps = {
+  title: 'Open',
+  isOpen: true,
+  onClick: () => {},
+};
+
+const OpenSectionheader = () => <SectionHeaderExampleTemplate title="Open Section Header" exampleProps={sectionHeaderProps} />;
+
+export default OpenSectionheader;