+++ packages/terra-tag/tests/nightwatch/DefaultTag.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Tag from '../../lib/Tag';
+
+const DefaultTag = () => <Tag id="default" text="Default Tag" />;
+
+export default DefaultTag;