+++ wdio.conf.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 const wdioConf = require('terra-toolkit/lib/wdio/conf');
-const WebpackDevService = require('terra-toolkit/lib/wdio/services/index').WebpackDevService;
+const ExpressDevService = require('terra-toolkit/lib/wdio/services/index').ExpressDevService;
 const localIP = require('ip');
 const path = require('path');
 const webpackConfig = require('./packages/terra-site/webpack.config.js');
@@ -28,7 +28,6 @@ const config = {
   // Ignore deprecation warnings. When chrome supports /actions API we'll update to use those.
   deprecationWarnings: false,
 
-  webpackPort,
   webpackConfig,
 
   axe: {
@@ -52,5 +51,5 @@ const config = {
   },
 };
 
-config.services = wdioConf.config.services.concat([WebpackDevService]);
+config.services = wdioConf.config.services.concat([ExpressDevService]);
 exports.config = config;