+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.3.0 - (March 14, 2018)
+------------------
+### Changed
+* Minor version bump
+
+2.2.0 - (March 6, 2018)
+------------------
+### Removed
+* Removed props-table script from package.json
+
 2.1.0 - (February 26, 2018)
 ------------------
 ### Changed